Finished Dual formulation of the problem and first dataset.
The class that defines implements the dual problem is finished, but seems that there is some bugs left in the code of the most violating pair.
On the other hand, the test with the first dataset (the breast_cancer_scale) give a good result with the LinearSVC model of scikit-learn, usefull for a comparison.
There is more to say about the dataset and how it was treated.